feat: console v2
took code out of program.cs into a seperate class and added ability to quit app whenever there is a input. Also added a method to get numbers from an input to be used for later.
